+++ package.json
@@ -25,5 +25,6 @@
   "engines": {
     "node": "16.x",
     "npm": "8.x"
-  }
+  },
+  "type": "module"
 }
